Issue #5, moved MockDataService to a separate file and added some documentation.
